Look for details in API documentation's changelog

+++ src/main/kotlin/eZmaxApi/models/SubnetMinusRequestCompound.kt
@@ -39,11 +39,11 @@ data class SubnetMinusRequestCompound (
 
     /* The network of the Subnet in integer form. For example 8.8.8.0 would be 134744064 */
     @Json(name = "iSubnetNetwork")
-    val iSubnetNetwork: kotlin.Int,
+    val iSubnetNetwork: kotlin.Long,
 
     /* The mask of the Subnet  in integer form. For example 255.255.255.0 would be 4294967040 */
     @Json(name = "iSubnetMask")
-    val iSubnetMask: kotlin.Int,
+    val iSubnetMask: kotlin.Long,
 
     /* The unique ID of the Subnet */
     @Json(name = "pkiSubnetID")
